Correccion de errores. ( Controladores, Service)

+++ src/main/java/com/example/Arevalo_Saibene_Nosicoski/Controller/ControladorDePaciente.java
@@ -34,7 +34,7 @@ public class ControladorDePaciente {
     }
 
     @GetMapping("/buscar/{id}")
-    public ResponseEntity<PacienteResponseDto> buscarPorId(@PathVariable Integer id){
+    public ResponseEntity<PacienteResponseDto> buscarPorId(@PathVariable Long id){
         PacienteResponseDto paciente = pacienteService.buscarPorId(id);
 
         if(paciente != null){
@@ -70,7 +70,7 @@ public class ControladorDePaciente {
     }
 
     @DeleteMapping("/eliminar/{id}")
-    public ResponseEntity<?> eliminarPaciente(@PathVariable Integer id)
+        public ResponseEntity<?> eliminarPaciente(@PathVariable Long id)
     {
         PacienteResponseDto pacienteEncontrado = pacienteService.buscarPorId(id);
